Move execution hooks to just before execution.
The execution hook was happening before the start of the decode
because there wasn't an easy way to inject after it. This meant that
it was not possible to report the instruction length. It might not
be a big problem, but it did mean that we couldn't call the hook
as easily. We now call the hook with the right details by having
a class that inherits from the Dispatcher do the hook calls.