Fix for BRK not dispatching to the IRQ handler.
The last address in the memory map regions would not be supplied - the
check for whether the memory was in the mapped object was not allowing
for the inclusive end address of the mapping. This meant that the
IRQ entry point (which is used by BRK) would be called with the high
address byte being returned from the standard memory, which was 0.
Instead of calling &DC1C, it was calling &001C.

We now check the memory region inclusively.