Fix PyBeebicorn's handling of register writes; hooks can now reset PC.
The register writes in the PyBeebicorn interface were always being
AND'd with 255, even for the PC. The value stored is checked in each
writer function now.

The execution hooks are now able to reset the PC value when they are
run, preventing the execution of the instruction that was decoded.