Add console handling - RDCH is able to work.
The RDCH entry point is replaced by a key reader from the console.
This uses the console code that I wrote for RISC OS Pyromaniac,
modified to be independant on Pyro.

This allows the RDCH entry point to be modified to handle the key
input. It's not great because we lose *Exec handling, but since we
don't have a filesystem or any other interfaces, that's not a big
deal right now.