Update the console.py module with full version from Pyromaniac.
I had originally cut down the console.py to be a limited, POSIX only,
version of the console.py implementation. However, there's no reason
why this cannot be used on Windows so I may as well include the
Windows version. It should work reasonably well, but it's not as
polished as the POSIX version.