Add code coverage for the tests that we have.
We can now exercise the code in some tests that can be run in a
Makefile. Only the coverage run is explicitly supported right now,
and it'll fail if the tests fail, so it's a combined coverage
and test reporter. We mix the unit tests and the integration tests
at the moment, and the output from the integration test isn't
checked for regressions.