Rework the base OS interfaces to be more flexible.
The base OS interfaces were duplicating some behaviour and not as
common as you might like. The interfaces are a little easier to
update now. It's not clear that they're actually going to make for
a better general experience but they will allow some more
flexibility.

Also a few typos fixed.