Add rudimetary host filesystem module and implementation of load/save.
We now have a host filesystem module, fsbbc.py, which provides a
BBC-like interface to the host filesystem. It's not tested beyond
being run through a few simple exercises, but the name translation
is based on code from RISC OS Pyromaniac so it should have the
right style of thing in it.

The OSFILE load and OSFILE save operations are implemented and these
allow BASIC to load files and save them.