Fixes for the Coverage runs.
Coverage runs of the Memory tests hadn't been working at all. They were
not invoking the unittest on the module. We now correctly invoke that
and get better coverage because of it.

The coverage had been confused on MacOS because PyBeeb.py and
pybeeb/__init__.py are effectively the same module, so we were getting
other references to modules that weren't really there. To avoid this
problem, and to make it easier to understand the difference between
the tool and the package, the command line program has been renamed
to `RunBeeb.py`.