Allow *Exec to work with the host readline; ctrl-c=>escape in RDCH.
When *Exec is used and we're trying to read a line, don't call the
host ReadLine interface. Instead we fall back to the standard
ReadLine in the MOS which will read characters and insert them into
the input from the file.

When ctrl-c is pressed whilst waiting for a key press, we now treat
this as an escape press.